+++ packages/tldraw/src/components/ToolsPanel/PrimaryTools.tsx
@@ -75,7 +75,7 @@ export const PrimaryTools = React.memo(function PrimaryTools(): JSX.Element {
       </ToolButtonWithTooltip>
       <ShapesMenu activeTool={activeTool} isToolLocked={isToolLocked} />
       <ToolButtonWithTooltip
-        kbd={'6'}
+        kbd={'7'}
         label={TDShapeType.Arrow}
         onClick={selectArrowTool}
         isLocked={isToolLocked}
@@ -84,7 +84,7 @@ export const PrimaryTools = React.memo(function PrimaryTools(): JSX.Element {
         <ArrowTopRightIcon />
       </ToolButtonWithTooltip>
       <ToolButtonWithTooltip
-        kbd={'7'}
+        kbd={'8'}
         label={TDShapeType.Text}
         onClick={selectTextTool}
         isLocked={isToolLocked}
@@ -93,7 +93,7 @@ export const PrimaryTools = React.memo(function PrimaryTools(): JSX.Element {
         <TextIcon />
       </ToolButtonWithTooltip>
       <ToolButtonWithTooltip
-        kbd={'8'}
+        kbd={'9'}
         label={TDShapeType.Sticky}
         onClick={selectStickyTool}
         isActive={activeTool === TDShapeType.Sticky}