+++ packages/tldraw/src/state/tools/LineTool/LineTool.ts
@@ -1,10 +1,10 @@
-import { Utils, TLPointerEventHandler, TLBoundsCorner } from '@tldraw/core'
-import { Ellipse } from '~state/shapes'
+import { Utils, TLPointerEventHandler } from '@tldraw/core'
+import { Arrow } from '~state/shapes'
 import { SessionType, TDShapeType } from '~types'
 import { BaseTool, Status } from '../BaseTool'
 
-export class EllipseTool extends BaseTool {
-  type = TDShapeType.Ellipse as const
+export class LineTool extends BaseTool {
+  type = TDShapeType.Line as const
 
   /* ----------------- Event Handlers ----------------- */
 
@@ -20,22 +20,21 @@ export class EllipseTool extends BaseTool {
 
     const id = Utils.uniqueId()
 
-    const newShape = Ellipse.create({
+    const newShape = Arrow.create({
       id,
       parentId: currentPageId,
       childIndex,
       point: currentPoint,
+      decorations: {
+        start: undefined,
+        end: undefined,
+      },
       style: { ...currentStyle },
     })
 
     this.app.patchCreate([newShape])
 
-    this.app.startSession(
-      SessionType.TransformSingle,
-      newShape.id,
-      TLBoundsCorner.BottomRight,
-      true
-    )
+    this.app.startSession(SessionType.Arrow, newShape.id, 'end', true)
 
     this.setStatus(Status.Creating)
   }