+++ packages/tldraw/src/types.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/ban-types */
 import type { TLPage, TLUser, TLPageState } from '@tldraw/core'
-import type { Command, Patch } from 'rko'
 import type { FileSystemHandle } from '~state/data/browser-fs-access'
 import type {
   TLBinding,
@@ -206,6 +205,7 @@ export type TDToolType =
   | TDShapeType.Draw
   | TDShapeType.Ellipse
   | TDShapeType.Rectangle
+  | TDShapeType.Line
   | TDShapeType.Arrow
   | TDShapeType.Sticky
 
@@ -271,6 +271,7 @@ export enum TDShapeType {
   Rectangle = 'rectangle',
   Draw = 'draw',
   Arrow = 'arrow',
+  Line = 'line',
   Text = 'text',
   Group = 'group',
 }
@@ -462,3 +463,11 @@ export type MappedByType<U extends string, T extends { type: U }> = {
 }
 
 export type ShapesWithProp<U> = MembersWithRequiredKey<MappedByType<TDShapeType, TDShape>, U>
+
+export type Patch<T> = Partial<{ [P in keyof T]: Patch<T[P]> }>
+
+export interface Command<T extends { [key: string]: any }> {
+  id?: string
+  before: Patch<T>
+  after: Patch<T>
+}